pyproject.toml
[tool.poetry]
name = "marcin-project"
version = "0.1.0"
description = ""
authors = ["mwojtyczka <marcin.wojtyczka@databricks.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.test.dependencies]
pyspark = "^3.5.0"
pytest = "^8.0.1"
pytest-cov = "^4.0.0"
pytest-spark = "^0.6.0"
chispa = "^0.9.2"
databricks-sdk = "^0.20.0"

[tool.poetry.group.int_test.dependencies]
pyspark = "^3.5.0"
pytest = "^8.0.1"
pytest-cov = "^4.0.0"
chispa = "^0.9.2"
databricks-connect = "^14.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "marcin_project.main:main"